+++ src/plugin-info.ts
@@ -1,5 +1,5 @@
 /* File will be overwritten by bin/release.sh! */
 export const PLUGIN_INFO = {
-  "pluginVersion": "1.5.0",
-  "pluginReleasedAt": "2024-02-19T13:37:05+0100"
+  "pluginVersion": "1.5.2",
+  "pluginReleasedAt": "2024-05-11T12:39:26+0200"
 }