+++ src/utils/zod.ts
@@ -1,7 +1,8 @@
 import { z } from "zod";
 import { TAbstractFile, TFile, TFolder } from "obsidian";
-import { activeVault } from "./file-handling";
+import { obsEnv } from "./obsidian-env";
 import { sanitizeFilePath } from "./file-handling";
+import { getEnabledCommunityPlugin, getEnabledCorePlugin } from "./plugins";
 
 // The absence of a parameter `blah`, a `blah=false` and a value-less `blah=`
 // should all be treated as `false`. My reign shall be merciful.
@@ -102,6 +103,28 @@ export const zodExistingNotePath = z.preprocess(
 
 /**
  * A schema which tests the passed-in string to see if it's a valid path to an
+ * existing template. If it is, returns a `TFile` instance.
+ */
+export const zodExistingTemplaterPath = z.preprocess(
+  lookupAbstractFileForTemplaterPath,
+  z.instanceof(TFile, {
+    message: "Template doesn't exist or Templater isn't enabled",
+  }),
+);
+
+/**
+ * A schema which tests the passed-in string to see if it's a valid path to an
+ * existing template. If it is, returns a `TFile` instance.
+ */
+export const zodExistingTemplatesPath = z.preprocess(
+  lookupAbstractFileForTemplatesPath,
+  z.instanceof(TFile, {
+    message: "Template doesn't exist or Templates isn't enabled",
+  }),
+);
+
+/**
+ * A schema which tests the passed-in string to see if it's a valid path to an
  * existing file. If it is, returns a `TFile` instance.
  */
 export const zodExistingFilePath = z.preprocess(
@@ -158,12 +181,9 @@ export const zodEmptyStringChangedToDefaultString = (defaultString: string) =>
  * @param path Any incoming zod parameter
  */
 function lookupAbstractFileForNotePath(path: any): TAbstractFile | null {
-  if (typeof path !== "string" || !path) {
-    return null;
-  }
-
-  const filepath = sanitizeFilePath(path as string);
-  return activeVault().getAbstractFileByPath(filepath);
+  return (typeof path === "string" && path.length > 0)
+    ? sanitizeFilePathAndGetAbstractFile(path)
+    : null;
 }
 
 /**
@@ -174,12 +194,9 @@ function lookupAbstractFileForNotePath(path: any): TAbstractFile | null {
  * @param path Any incoming zod parameter
  */
 function lookupAbstractFileForFilePath(path: any): TAbstractFile | null {
-  if (typeof path !== "string" || !path) {
-    return null;
-  }
-
-  const filepath = sanitizeFilePath(path as string, false);
-  return activeVault().getAbstractFileByPath(filepath);
+  return (typeof path === "string" && path.length > 0)
+    ? sanitizeFilePathAndGetAbstractFile(path, false)
+    : null;
 }
 
 /**
@@ -190,9 +207,73 @@ function lookupAbstractFileForFilePath(path: any): TAbstractFile | null {
  * @param path Any incoming zod parameter
  */
 function lookupAbstractFolderForPath(path: any): TAbstractFile | null {
+  return (typeof path === "string" && path.length > 0)
+    ? obsEnv.activeVault.getAbstractFileByPath(path as string)
+    : null;
+}
+
+function sanitizeFilePathAndGetAbstractFile(
+  path: string,
+  isNote?: boolean,
+): TAbstractFile | null {
+  return obsEnv.activeVault
+    .getAbstractFileByPath(sanitizeFilePath(path, isNote));
+}
+
+/**
+ * Takes an incoming parameter and returns the corresponding `TAbstractFile` if
+ * the parameter is a string and the string corresponds to an existing template
+ * file. If the passed in path can't be found, the function will also check
+ * Templater's template folder path for the file. Returns `null` when the search
+ * came up empty.
+ *
+ * @param path Any incoming zod parameter
+ * @returns
+ */
+function lookupAbstractFileForTemplaterPath(path: any): TAbstractFile | null {
   if (typeof path !== "string" || !path) {
     return null;
   }
 
-  return activeVault().getAbstractFileByPath(path as string);
+  const abstractFile = sanitizeFilePathAndGetAbstractFile(path, false) ||
+    sanitizeFilePathAndGetAbstractFile(`${path}.md`, false);
+  if (abstractFile) return abstractFile;
+
+  const res = getEnabledCommunityPlugin("templater-obsidian");
+  if (res.isSuccess) {
+    const folder = res.result.settings?.templates_folder;
+    return sanitizeFilePathAndGetAbstractFile(`${folder}/${path}`, false) ||
+      sanitizeFilePathAndGetAbstractFile(`${folder}/${path}.md`, false);
+  }
+
+  return null;
+}
+
+/**
+ * Takes an incoming parameter and returns the corresponding `TAbstractFile` if
+ * the parameter is a string and the string corresponds to an existing template
+ * file. If the passed in path can't be found, the function will also check
+ * Templates' template folder path for the file. Returns `null` when the search
+ * came up empty.
+ *
+ * @param path Any incoming zod parameter
+ * @returns
+ */
+function lookupAbstractFileForTemplatesPath(path: any): TAbstractFile | null {
+  if (typeof path !== "string" || !path) {
+    return null;
+  }
+
+  const abstractFile = sanitizeFilePathAndGetAbstractFile(path, false) ||
+    sanitizeFilePathAndGetAbstractFile(`${path}.md`, false);
+  if (abstractFile) return abstractFile;
+
+  const res = getEnabledCorePlugin("templates");
+  if (res.isSuccess) {
+    const folder = res.result.options?.folder;
+    return sanitizeFilePathAndGetAbstractFile(`${folder}/${path}`, false) ||
+      sanitizeFilePathAndGetAbstractFile(`${folder}/${path}.md`, false);
+  }
+
+  return null;
 }